+++ src/clj/rems/db/entitlements.clj
@@ -3,7 +3,6 @@
   (:require [clj-http.client :as http]
             [clj-time.core :as time]
             [clojure.set :refer [union]]
-            [clojure.string :refer [join]]
             [clojure.tools.logging :as log]
             [mount.core :as mount]
             [rems.application-util :as application-util]
@@ -11,6 +10,7 @@
             [rems.config :refer [env]]
             [rems.db.applications :as applications]
             [rems.db.core :as db]
+            [rems.db.csv :as csv]
             [rems.db.outbox :as outbox]
             [rems.db.users :as users]
             [rems.json :as json]
@@ -42,12 +42,8 @@
   []
   (when-not (has-roles? :handler)
     (throw-forbidden))
-  (let [ents (db/get-entitlements)
-        separator (:csv-separator env)]
-    (with-out-str
-      (println (join separator ["resource" "application" "user" "start"]))
-      (doseq [e ents]
-        (println (join separator [(:resid e) (:catappid e) (:userid e) (text/localize-time (:start e))]))))))
+  (let [ents (db/get-entitlements)]
+    (csv/entitlements-to-csv ents)))
 
 (defn- post-entitlements! [{:keys [entitlements action] :as params}]
   (when-let [target (get-in env [:entitlements-target action])]